requirements.txt
# Core dependencies
openai>=1.3.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0

# PDF processing
PyMuPDF>=1.23.0
pdfplumber>=0.7.0

# NLP and similarity
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Academic APIs
arxiv>=1.4.0

# Text processing
nltk>=3.8.0
spacy>=3.6.0

# Optional: Graph processing
opencv-python>=4.8.0
matplotlib>=3.7.0
plotly>=5.15.0
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Build tools for Python 3.13 compatibility
setuptools>=68.0.0
wheel>=0.41.0
